Attempt to access 401 unauthorised via DocumentService signs user in
If Angular detects a location change and the DocumentService is unauthorised to access the information, the user is automatically redirected to the Omniauth Google strategy, passing the restricted page as the 'continue' param; if successful, the user will automatically be logged back in and access will be provided.

This only occurs if the user is able to click a link to a restricted page from inside the Angular app; such as when the user has two tabs open and logs out in one and then tries to access a restricted page in the other.